Refactor (#181)
* fixed create-ip for 7 tuple drs

* brought cmip6 up to date

* bring in handle subaxis for long

* dataset aggregator added

* skip prepare and change toward mkd class

* fixes

* updated mapfile handling

* updated mapfile handling to include multiple dirs

* cmip5 in progress, update fixed

* cmip5 changes

* cmip5 publishes, variable settings todo

* cmip5 working, create-ip aggregate dataset failing on create-ip

* fixed mapfile input issues

* updated format_template

* updated format_template again

* updated pub_client to exit with correct code on failure

* current create-ip progress

* create-ip list copy issue still persisting

* Fixed create-ip issue

* cleaned up debugging print lines

* removed another print line

* fixed cmip5 autocurator and specified variable finding

* initial

* refactor for input4mips

* add input4mips

* fix indent for projs other than cmip6

* add skip_prepare to args

* input4mips working

* made project case insensitive

* added handling of user defined project to pub_internal

* esgmigrate working command line

* user defined projects work

* fix the xattr

* e3sm initial commit

* reduce urls

* e3sm working

* reduce urls 2nd attempt

* correct the xattr refactor, add stack trace, url filter None

* updated non-nc

* docs in progress, modules refactored, yet to be tested

* updated last of docs

* updated last of docs

* Update README.md

* changed spacing

* fixed individual modules

* Update esgindexpub.rst

* fixed update docs

* refactored init using super

* fixed init

* correct settings typos, add error return

* xattr bugfix

* made settings case insensitive

* added esgunpublish to docs

* specified certain behaviors

* fixed custom projects

* fixed xattr handling

* clarified ini file errors

* temporary variable issue fix

* updating docs

* updated esgmigrate cmd

* updated esgmigrate cli args

* made changes to esgmigrate, program not working

* fixes for esgmigrate

* fixed esgmigrate.py

* add an intro and TLDR

* fix TLDR and intro points

* remove line break

+++ pkg/esgcet/esgmigrate.py
@@ -7,7 +7,7 @@ from datetime import date
 from pathlib import Path
 import json
 
-DEFAULT_ESGINI = '/esg/config/esgcet'
+DEFAULT_ESGINI = '/esg/config/esgcet/'
 CONFIG_FN_DEST = "~/.esg/esg.ini"
 
 
@@ -17,10 +17,11 @@ def project_list(cfg_obj):
 
 class ESGPubMigrate(object):
 
-    def __init__(self, args):
-        self.ini_path = args[0]
-        self.silent = args[1]
-        self.verbose = args[2]
+    def __init__(self, i, newpath, silent=False, verbose=False):
+        self.ini_path = i
+        self.silent = silent
+        self.verbose = verbose
+        self.save_path = newpath
 
     def project_migrate(self, project):
 
@@ -43,12 +44,8 @@ class ESGPubMigrate(object):
 
     def migrate(self, project=None):
 
-        if os.path.exists(CONFIG_FN_DEST):
-            print("New config file exists, exiting.", file=sys.stderr)
-            exit(0)
-
-        if not os.path.exists(self.ini_path + '/esg.ini'):
-            print("esg.ini not found or unreadable", file=sys.stderr)
+        if not os.path.exists(self.ini_path + "esg.ini"):
+            print("Old config " + self.ini_path + "esg.ini not found or unreadable.", file=sys.stderr)
             exit(1)
 
         try:
@@ -97,8 +94,8 @@ class ESGPubMigrate(object):
         except:
             svc_urls = []
 
-        DATA_TRANSFER_NODE = ""
-        GLOBUS_UUID = ""
+        DATA_TRANSFER_NODE = None
+        GLOBUS_UUID = None
 
         for line in svc_urls:
             if line[0] == "GridFTP":
@@ -120,27 +117,35 @@ class ESGPubMigrate(object):
             print(CERT_FN)
             print(DATA_TRANSFER_NODE)
             print(GLOBUS_UUID)
+            print(project)
 
         project_config = {project: self.project_migrate(project)}
 
         d = date.today()
         t = d.strftime("%y%m%d")
         home = str(Path.home())
-        config_file = home + "/.esg/esg.ini"
-        backup = home + "/.esg/" + t + "esg.ini"
-        shutil.copyfile(config_file, backup)
+        config_file = self.save_path
+        if os.path.exists(self.save_path):
+            backup = self.save_path + ".bak"
+            shutil.copyfile(config_file, backup)
+        Path(config_file).touch()
         config = cfg.ConfigParser()
         config.read(config_file)
         new_config = {"data_node": data_node, "index_node": index_node, "data_roots": json.dumps(dr_dict), "cert": CERT_FN,
                       "globus_uuid": GLOBUS_UUID, "data_transfer_node": DATA_TRANSFER_NODE, "pid_creds": json.dumps(pid_creds)}
 
-        if project_config:
-            new_config["project_cfg"] = json.dumps(project_config)
-        for key, value in new_config.items():
-            try:
-                test = config['user'][key]
-            except:
-                config['user'][key] = value
+        try:
+            test = config['user']['data_node']
+            if project_config:
+                new_config["project_cfg"] = json.dumps(project_config)
+            for key, value in new_config.items():
+                if value:
+                    config.set('user', key, value)
+        except:
+            config.add_section('user')
+            for key, value in new_config.items():
+                if value:
+                    config.set('user', key, value)
         with open(config_file, "w") as cf:
             config.write(cf)
